Add 'native_kpsewhich' feature
Previously this project required the 'libkpathsea' library in order to
compile properly. This is not an issue on sufficiently new versions of
Ubuntu, but was a problem on OS X machines.

Instead of using kpsewhich form the kpathsea library, it is possible to
use the 'kpsewhich' executable at runtime. This commit implements this
approach and hides the previous approach behind a 'native_kpsewhich'
feature flag.